requirements.txt
streamlit==1.33.0
selenium==4.21.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
langchain==0.2.1
langchain-nvidia-ai-endpoints==0.1.7
transformers==4.41.2